# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Budget Monitor API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (admin only)
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserWithSector"

    post:
      operationId: createUser
      tags: [users]
      summary: Create a new user (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserWithSector"
  /users/{userId}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user by ID
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserWithSector"

    put:
      operationId: updateUser
      tags: [users]
      summary: Update user (admin only)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserWithSector"

    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user (admin only)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # SECTORS
  /sectors:
    get:
      operationId: listSectors
      tags: [sectors]
      summary: List all sectors visible to the current user
      responses:
        "200":
          description: List of sectors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SectorWithStats"

    post:
      operationId: createSector
      tags: [sectors]
      summary: Create a new sector (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSectorBody"
      responses:
        "201":
          description: Sector created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sector"
  /sectors/tree:
    get:
      operationId: getSectorTree
      tags: [sectors]
      summary: Get full hierarchical sector tree with budget balances
      parameters:
        - name: cycleId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Hierarchical tree of sectors with balances
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SectorTreeNode"
  /sectors/{sectorId}:
    get:
      operationId: getSector
      tags: [sectors]
      summary: Get a sector by ID
      parameters:
        - name: sectorId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Sector details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SectorWithStats"

    put:
      operationId: updateSector
      tags: [sectors]
      summary: Update a sector (admin only)
      parameters:
        - name: sectorId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSectorBody"
      responses:
        "200":
          description: Sector updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sector"

    delete:
      operationId: deleteSector
      tags: [sectors]
      summary: Delete a sector (admin only)
      parameters:
        - name: sectorId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Sector deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /sectors/{sectorId}/tree:
    get:
      operationId: getSectorSubtree
      tags: [sectors]
      summary: Get tree for a specific sector with budget balances
      parameters:
        - name: sectorId
          in: path
          required: true
          schema:
            type: integer
        - name: cycleId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Sector subtree with balances
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SectorTreeNode"

  # BUDGET CYCLES
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [super_admin, ceo, ministry_head, department_head, viewer]
        sectorId:
          type: integer
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - role
        - isActive
        - createdAt
    UserWithSector:
      allOf:
        - $ref: "#/components/schemas/User"
        - type: object
          properties:
            sector:
              $ref: "#/components/schemas/Sector"
              nullable: true
    CreateUserBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
        role:
          type: string
          enum: [super_admin, ceo, ministry_head, department_head, viewer]
        sectorId:
          type: integer
          nullable: true
      required:
        - name
        - email
        - password
        - role
    UpdateUserBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum: [super_admin, ceo, ministry_head, department_head, viewer]
        sectorId:
          type: integer
          nullable: true
        isActive:
          type: boolean
        password:
          type: string
    Sector:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        code:
          type: string
        parentId:
          type: integer
          nullable: true
        depth:
          type: integer
        responsibleUserId:
          type: integer
          nullable: true
        isActive:
          type: boolean
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - code
        - depth
        - isActive
        - sortOrder
        - createdAt
    SectorWithStats:
      allOf:
        - $ref: "#/components/schemas/Sector"
        - type: object
          properties:
            responsibleUser:
              $ref: "#/components/schemas/User"
              nullable: true
            parent:
              $ref: "#/components/schemas/Sector"
              nullable: true
            totalAllocated:
              type: number
            totalRevoked:
              type: number
            netAllocated:
              type: number
            availableBalance:
              type: number
            utilizationPct:
              type: number
            childCount:
              type: integer
    SectorTreeNode:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        code:
          type: string
        depth:
          type: integer
        parentId:
          type: integer
          nullable: true
        totalAllocated:
          type: number
        totalRevoked:
          type: number
        netAllocated:
          type: number
        availableBalance:
          type: number
        utilizationPct:
          type: number
        responsibleUser:
          $ref: "#/components/schemas/User"
          nullable: true
        children:
          type: array
          items:
            $ref: "#/components/schemas/SectorTreeNode"
      required:
        - id
        - name
        - code
        - depth
        - totalAllocated
        - totalRevoked
        - netAllocated
        - availableBalance
        - utilizationPct
        - children
    CreateSectorBody:
      type: object
      properties:
        name:
          type: string
        code:
          type: string
        parentId:
          type: integer
          nullable: true
        responsibleUserId:
          type: integer
          nullable: true
        sortOrder:
          type: integer
      required:
        - name
        - code
    UpdateSectorBody:
      type: object
      properties:
        name:
          type: string
        code:
          type: string
        parentId:
          type: integer
          nullable: true
        responsibleUserId:
          type: integer
          nullable: true
        isActive:
          type: boolean
        sortOrder:
          type: integer
...